+++ src/lib/core/tests/TestCHIPTLV.cpp
@@ -2565,13 +2565,13 @@ void CheckCHIPTLVPutStringF(nlTestSuite * inSuite, void * inContext)
     uint8_t backingStore[bufsize];
     TLVWriter writer;
     TLVReader reader;
-    unsigned int num = 1;
-    CHIP_ERROR err   = CHIP_NO_ERROR;
+    size_t num     = 1;
+    CHIP_ERROR err = CHIP_NO_ERROR;
 
     writer.Init(backingStore, bufsize);
-    snprintf(strBuffer, sizeof(strBuffer), "Sample string %u", num);
+    snprintf(strBuffer, sizeof(strBuffer), "Sample string %u", static_cast<unsigned int> num);
 
-    err = writer.PutStringF(ProfileTag(TestProfile_1, 1), "Sample string %u", num);
+    err = writer.PutStringF(ProfileTag(TestProfile_1, 1), "Sample string %u", static_cast<unsigned int> num);
     NL_TEST_ASSERT(inSuite, err == CHIP_NO_ERROR);
 
     err = writer.Finalize();
@@ -2645,13 +2645,13 @@ void CheckCHIPTLVPutStringFCircular(nlTestSuite * inSuite, void * inContext)
     CircularTLVWriter writer;
     CircularTLVReader reader;
     CHIPCircularTLVBuffer buffer(backingStore, bufsize);
-    unsigned int num = 1;
-    CHIP_ERROR err   = CHIP_NO_ERROR;
+    size_t num     = 1;
+    CHIP_ERROR err = CHIP_NO_ERROR;
 
     // Initial test: Verify that a straight printf works as expected into continuous buffer.
 
     writer.Init(buffer);
-    snprintf(strBuffer, sizeof(strBuffer), "Sample string %u", num);
+    snprintf(strBuffer, sizeof(strBuffer), "Sample string %u", static_cast<unsigned int> num);
 
     err = writer.PutBoolean(ProfileTag(TestProfile_1, 2), true);
     NL_TEST_ASSERT(inSuite, err == CHIP_NO_ERROR);
@@ -2659,7 +2659,7 @@ void CheckCHIPTLVPutStringFCircular(nlTestSuite * inSuite, void * inContext)
     err = writer.Finalize();
     NL_TEST_ASSERT(inSuite, err == CHIP_NO_ERROR);
 
-    err = writer.PutStringF(ProfileTag(TestProfile_1, 1), "Sample string %u", num);
+    err = writer.PutStringF(ProfileTag(TestProfile_1, 1), "Sample string %u", static_cast<unsigned int> num);
     NL_TEST_ASSERT(inSuite, err == CHIP_NO_ERROR);
 
     err = writer.Finalize();
@@ -2683,9 +2683,9 @@ void CheckCHIPTLVPutStringFCircular(nlTestSuite * inSuite, void * inContext)
     // This print will both stradle the boundary of the buffer and displace the previous two elements.
     num = 2;
 
-    snprintf(strBuffer, sizeof(strBuffer), "Sample string %u", num);
+    snprintf(strBuffer, sizeof(strBuffer), "Sample string %u", static_cast<unsigned int> num);
 
-    err = writer.PutStringF(ProfileTag(TestProfile_1, 1), "Sample string %u", num);
+    err = writer.PutStringF(ProfileTag(TestProfile_1, 1), "Sample string %u", static_cast<unsigned int> num);
     NL_TEST_ASSERT(inSuite, err == CHIP_NO_ERROR);
 
     err = writer.Finalize();
